fix some bugs about sql_type
