+++ pkg/sql/colexec/external/external_test.go
@@ -26,6 +26,7 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/container/types"
 	"github.com/matrixorigin/matrixone/pkg/sql/parsers/tree"
 	"github.com/matrixorigin/matrixone/pkg/sql/plan"
+	plan2 "github.com/matrixorigin/matrixone/pkg/sql/plan"
 	"github.com/matrixorigin/matrixone/pkg/testutil"
 	"github.com/matrixorigin/matrixone/pkg/vm/process"
 	"github.com/smartystreets/goconvey/convey"
@@ -575,7 +576,7 @@ func TestReadDirSymlink(t *testing.T) {
 
 	// read a/b/d/foo
 	fooPathInB := filepath.Join(root, "a", "b", "d", "foo")
-	files, err := ReadDir(&tree.ExternParam{
+	files, err := plan2.ReadDir(&tree.ExternParam{
 		Filepath: fooPathInB,
 		Ctx:      ctx,
 	})
@@ -584,7 +585,7 @@ func TestReadDirSymlink(t *testing.T) {
 	assert.Equal(t, fooPathInB, files[0])
 
 	path1 := root + "/a//b/./../b/c/foo"
-	files1, err := ReadDir(&tree.ExternParam{
+	files1, err := plan2.ReadDir(&tree.ExternParam{
 		Filepath: path1,
 		Ctx:      ctx,
 	})
