+++ test/distributed/cases/sql_source_type/sql_source_type.result
@@ -1,5 +1,5 @@
 /* cloud_user */drop table if exists __mo_t1;
-/* cloud_nouser */ create table __mo_t1(a int);
+/* cloud_nonuser */ create table __mo_t1(a int);
 insert into __mo_t1 values(1);
 select * from __mo_t1;
 a
@@ -8,12 +8,10 @@ use system;
 select sleep(15);
 sleep(15)
 0
-select statement, sql_source_type from statement_info where sql_source_type = 'cloud_user_sql' limit 1;
+select statement, sql_source_type from statement_info where user="dump" order by request_at desc limit 5;
 statement    sql_source_type
-drop table if exists __mo_t1    cloud_user_sql
-select statement, sql_source_type from statement_info where sql_source_type = 'cloud_nouser_sql' limit 1;
-statement    sql_source_type
-create table __mo_t1 (a int)    cloud_nouser_sql
-select statement, sql_source_type from statement_info where statement = 'insert into __mo_t1 values (1)' limit 1;
-statement    sql_source_type
+select sleep(15)    external_sql
+use system    external_sql
+select * from __mo_t1    external_sql
 insert into __mo_t1 values (1)    external_sql
+create table __mo_t1 (a int)    cloud_nonuser_sql
